linkedql: move main types to a single file
